+++ diesel/src/macros/ops.rs
@@ -63,3 +63,21 @@ macro_rules! __diesel_generate_ops_impls_if_numeric {
 
     ($column_name:ident, $non_numeric_type:ty) => {};
 }
+
+#[macro_export]
+#[doc(hidden)]
+macro_rules! date_time_expr {
+    ($tpe: ty) => {
+        operator_allowed!($tpe, Add, add);
+        operator_allowed!($tpe, Sub, sub);
+    }
+}
+
+#[macro_export]
+#[doc(hidden)]
+macro_rules! __diesel_generate_ops_impls_if_date_time {
+    ($column_name:ident, Nullable<$($inner:tt)::*>) => { __diesel_generate_ops_impls_if_date_time!($column_name, $($inner)::*); };
+    ($column_name:ident, Date) => { date_time_expr!($column_name); };
+    ($column_name:ident, Timestamp) => { date_time_expr!($column_name); };
+    ($column_name:ident, $non_date_time_type:ty) => {};
+}